Add tap gesture recognizer to ListItemTableView and create custom cell
Add tap gesture recognizer to mimic the Add functionality of the
Reminders app. When a user taps the table and not on a cell, the
next available cell in the list should be visible, editable and
becomes first responder. This is one of several nuance taps that
replace an explicit Add button.

When a user taps on the next available line, this should also
Add a record that can be edited. This however, will be implemented
in a different code set than the tap gesture recognizer on the
ListItemTableView.